complete/requirements.txt
selenium
mysql-connector-python
pyinstaller